Replace database with in-memory variables
The data is now stored in two variables; 'all_users' and
'all_messages'.

The user_id is no longer stored alongside the other user data -
it's the key. This makes it a bit awkward to pass to the template;
either it needs to be merged with the user data or passed separately.
To avoid this problem, the template now uses the username in place
of the user_id, because it works just as well (the username is also
guaranteed to be unique).

Augmenting the messages to contain user information was a bit
tedious. It was done with a table join previously; instead
a helper function is now used. I hope to remove the need for this
in the near future.